requirements.txt
flask==3.1.2
python-dotenv==1.0.1
pandas==3.0.1
openpyxl==3.1.5
reportlab==4.4.10
waitress==3.0.2